Re-arrangement of tables
