+++ .github/pull_request_template.md
@@ -19,3 +19,9 @@ The below needs to be done before a pull request can be approved:
 - [ ] New components added as an export to src/index.js
 - [ ] New components/features have sufficient unit tests
 - [ ] New components/features have sufficient documentation
+
+## Accesibility check list
+
+- [ ] New components/features are accessible to keyboard users (All links/buttons are tabbable, All content is accessible)
+- [ ] New components/features are accessible to non-JS users (All links/buttons are tabbable, All content is accessible), this may have visual differences
+- [ ] New components/features have hover, focus and active states on all the links/buttons
