+++ assets/sass/_functions/variables.scss
@@ -89,7 +89,7 @@ $dark:          #46003c;
 // Functional/state variables 
 $colour-body: #595959;
 $colour-muted: #9d9d9d;
-$colour-border: #6f6f6f;
+$colour-border: #D8D8D8;
 $colour-link: $primary;
 $colour-focus: #000000;
 $colour-hover: #000000;
