+++ assets/ts/components/table/README.md
@@ -1,0 +1,13 @@
+```
+<iam-table class="container">
+  <table>
+    <thead>...</thead>
+    <tbody>...</tbody>
+  </table>
+</iam-tabs>
+```
+
+### Properties
+
+| Option | Type | Default Value | Description |
+| ------ | ---- | ------------- | ----------- |
