+++ docs/routes.ts
@@ -294,14 +294,6 @@ const routes = [
         component: () => import('./views/components/Tables.vue')
       },
       {
-        path: 'advanced-tables',
-        name: 'Advanced Tables',
-        meta: {
-          title: 'Advanced Tables | Elements | iamkey'
-        },
-        component: () => import('./views/components/AdvancedTables.vue')
-      },
-      {
         path: 'tooltips',
         name: 'Tooltips',
         meta: {
