+++ flat-components.html
@@ -120,6 +120,68 @@
         </details>
       </div>
       
+
+
+      <div class="container">
+        <div class="table--sticky" style="--cta-width:8rem;">
+        <div class="table__wrapper table--minify">
+          <table>
+            <thead>
+              <tr>
+                <th>Client</th>
+                <th>Date added</th>
+                <th>Transaction address</th>
+                <th>Onboarding status</th>
+                <th>Risk level</th>
+                <th>Due diligence</th>
+                <th></th>
+              </tr>
+            </thead>
+            <tbody>
+              <tr>
+                <td>Dave Smith, Buyer</td>
+                <td>18 Mar 23</td>
+                <td>89, Queens Crescent, NE28 8DR</td>
+                <td>Risk assessment complete</td>
+                <td>Low</td>
+                <td>Verified</td>
+                <td><a href="/clienturl" class="text-nowrap">Manage client</a></td>  
+              </tr>
+              <tr>
+                <td>Dave Smith, Buyer</td>
+                <td>18 Mar 23</td>
+                <td>89, Queens Crescent, NE28 8DR</td>
+                <td>Risk assessment complete</td>
+                <td>Low</td>
+                <td><a href="/clienturl" class="text-nowrap">Verified</a></td>
+                <td><a href="/clienturl" class="text-nowrap">Manage client</a></td>  
+              </tr>
+              <tr>
+                <td>Dave Smith, Buyer</td>
+                <td>18 Mar 23</td>
+                <td>89, Queens Crescent, NE28 8DR</td>
+                <td>Risk assessment complete</td>
+                <td>Low</td>
+                <td>Verified</td>
+                <td><a href="/clienturl" class="text-nowrap">Manage client</a></td>  
+              </tr>
+              <tr>
+                <td>Dave Smith, Buyer</td>
+                <td>18 Mar 23</td>
+                <td>89, Queens Crescent, NE28 8DR</td>
+                <td>Risk assessment complete</td>
+                <td>Low</td>
+                <td>Verified</td>
+                <td><a href="/clienturl" class="text-nowrap">Manage client</a></td>  
+              </tr>
+              
+            </tbody>
+          </table>
+        </div>
+        </div>
+      </div>
+
+
     </main>
 
   </body>
